+++ blocks/form/src/index.tsx
@@ -34,6 +34,10 @@ const validators: { [name: string]: Validator } = {
       return true;
     }
 
+    if (value === null) {
+      return false;
+    }
+
     if (field.accept) {
       if (field.repeated) {
         return (